Move marker components to separate files
